+++ Util.h
@@ -2,13 +2,14 @@
 #include<cuda_runtime_api.h>
 #include<device_functions.h>
 #include<device_launch_parameters.h>
+#include<curand_kernel.h>
 #include<limits>
 #include<stdio.h>
 
-__constant__ const double infinity = std::numeric_limits<double>::infinity();
-__constant__ const double pi = 3.1415926535897932385;
+__constant__ const float infinity = std::numeric_limits<float>::infinity();
+__constant__ const float pi = 3.1415926535897932385;
 
-inline double Deg2Rad(double degrees) {
+inline float Deg2Rad(float degrees) {
 	return degrees * pi / 180.0;
 }
